+++ HISTORY.md
@@ -37,6 +37,9 @@
     * German
     * Chinese, thanks, Simon!
 
+### 2021-10-06
+* threads: programmatically hide individual variables in palette
+
 ### 2021-10-05
 * threads, store: added infrastructure for hiding individual variables in palette